Extract SettingsCommonScreen
