+++ design/src/main/java/com/github/kr328/clash/design/component/SettingsCommonCompose.kt
@@ -1,0 +1,131 @@
+package com.github.kr328.clash.design.component
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.ColumnScope
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.WindowInsets
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import com.github.kr328.clash.design.R
+
+@Composable
+@OptIn(ExperimentalMaterial3Api::class)
+fun SettingsCommonScreen(
+  title: String,
+  onBack: () -> Unit,
+  modifier: Modifier = Modifier,
+  content: @Composable ColumnScope.() -> Unit,
+) {
+  val scrollBehavior = TopAppBarDefaults.pinnedScrollBehavior()
+
+  Scaffold(
+    modifier = modifier,
+    topBar = {
+      TopAppBar(
+        title = { Text(text = title, maxLines = 1, overflow = TextOverflow.Ellipsis) },
+        navigationIcon = {
+          IconButton(onClick = onBack) {
+            Icon(
+              painter = painterResource(id = R.drawable.ic_baseline_arrow_back),
+              contentDescription = stringResource(id = R.string.close),
+            )
+          }
+        },
+        scrollBehavior = scrollBehavior,
+      )
+    },
+    contentWindowInsets = WindowInsets(0),
+  ) { innerPadding ->
+    Column(
+      modifier = Modifier.fillMaxSize().padding(innerPadding).verticalScroll(rememberScrollState())
+    ) {
+      content()
+    }
+  }
+}
+
+@Composable
+fun SettingsTipsItem(text: String, modifier: Modifier = Modifier) {
+  Row(
+    modifier = modifier.padding(vertical = 16.dp),
+    verticalAlignment = Alignment.CenterVertically,
+  ) {
+    Spacer(modifier = Modifier.width(20.dp))
+
+    Image(
+      painter = painterResource(id = R.drawable.ic_outline_info),
+      contentDescription = null,
+      modifier = Modifier.size(25.dp),
+    )
+
+    Spacer(modifier = Modifier.width(20.dp))
+
+    Text(
+      text = text,
+      style = MaterialTheme.typography.bodyMedium,
+      modifier = Modifier.padding(end = 20.dp),
+    )
+  }
+}
+
+@Composable
+fun SettingsCategoryTitle(text: String, modifier: Modifier = Modifier) {
+  Text(
+    text = text,
+    style = MaterialTheme.typography.bodyMedium,
+    color = MaterialTheme.colorScheme.primary,
+    modifier =
+      modifier.fillMaxWidth().padding(start = 65.dp, end = 20.dp, top = 16.dp, bottom = 16.dp),
+  )
+}
+
+@Composable
+fun SettingsClickableItem(
+  title: String,
+  summary: String,
+  onClick: () -> Unit,
+  modifier: Modifier = Modifier,
+) {
+  Row(
+    modifier =
+      modifier
+        .fillMaxWidth()
+        .clickable(onClick = onClick)
+        .padding(top = 16.dp, bottom = 16.dp, end = 20.dp),
+    verticalAlignment = Alignment.CenterVertically,
+  ) {
+    Spacer(modifier = Modifier.width(65.dp))
+
+    Column {
+      Text(text = title, style = MaterialTheme.typography.bodyLarge)
+      Text(
+        text = summary,
+        style = MaterialTheme.typography.bodyMedium,
+        modifier = Modifier.padding(top = 5.dp),
+      )
+    }
+  }
+}
